refactor: 파일 구조 변경
JIRA: GUC-113
resolves: #11
