+++ src/main/java/com/gucci/message_service/domain/Message.java
@@ -58,8 +58,4 @@ public class Message {
         this.deletedBySender = true;
     }
 
-    public void markAsRead() {
-        this.isRead = true;
-    }
-
 }
